chore: move winston to maintained packages
