--- deprecated/core-logger-winston/README.md
@@ -1,26 +1,0 @@
-# Ark Core - Winston Logger
-
-<p align="center">
-    <img src="https://raw.githubusercontent.com/ArkEcosystem/core/master/banner.png?sanitize=true" />
-</p>
-
-## Deprecated
-Note that this plugin is deprecated and should no longer be used
-
-## Documentation
-
-You can find installation instructions and detailed instructions on how to use this package at the [dedicated documentation site](https://docs.ark.io/guidebook/core/plugins/deprecated/core-logger-winston.html).
-
-## Security
-
-If you discover a security vulnerability within this package, please send an e-mail to security@ark.io. All security vulnerabilities will be promptly addressed.
-
-## Credits
-
--   [Brian Faust](https://github.com/faustbrian)
--   [François-Xavier Thoorens](https://github.com/fix)
--   [All Contributors](../../../../contributors)
-
-## License
-
-[MIT](LICENSE) © [ArkEcosystem](https://ark.io)
