+++ packages/core-logger-winston/.gitattributes
