+++ packages/core-logger-winston/README.md
@@ -1,0 +1,21 @@
+# Ark Core - Winston Logger
+
+<p align="center">
+    <img src="https://raw.githubusercontent.com/ArkEcosystem/core/master/banner.png?sanitize=true" />
+</p>
+
+## Documentation
+
+You can find installation instructions and detailed instructions on how to use this package at the [dedicated documentation site](https://docs.ark.io/guidebook/core/plugins/optional/core-logger-winston.html).
+
+## Security
+
+If you discover a security vulnerability within this package, please send an e-mail to security@ark.io. All security vulnerabilities will be promptly addressed.
+
+## Credits
+
+This project exists thanks to all the people who [contribute](../../../../contributors).
+
+## License
+
+[MIT](LICENSE) © [ArkEcosystem](https://ark.io)
