+++ packages/core-logger-winston/package.json
@@ -1,7 +1,7 @@
 {
     "name": "@arkecosystem/core-logger-winston",
     "description": "Winston Logger for Ark Core",
-    "version": "2.2.0",
+    "version": "2.3.0-next.0",
     "contributors": [
         "François-Xavier Thoorens <fx@ark.io>",
         "Brian Faust <brian@ark.io>"
@@ -25,8 +25,8 @@
         "test:watch:all": "cross-env CORE_ENV=test jest --runInBand --watchAll"
     },
     "dependencies": {
-        "@arkecosystem/core-interfaces": "^2.2.0",
-        "@arkecosystem/core-logger": "^2.2.0",
+        "@arkecosystem/core-interfaces": "^2.3.0-next.0",
+        "@arkecosystem/core-logger": "^2.3.0-next.0",
         "chalk": "^2.4.2",
         "colors": "^1.3.3",
         "dayjs-ext": "^2.2.0",
