+++ packages/core-logger-winston/src/defaults.ts
