+++ packages/core-logger-winston/src/formatter.ts
