+++ packages/core-logger-winston/src/plugin.ts
